Mailer.php

+++ Mailer.php
@@ -458,21 +458,18 @@ class Pman_Core_Mailer {
         $user = $uinfo['name']; 
         
         $cache = ini_get('session.save_path')."/Pman_Core_Mailer-{$user}/" . md5($url);
-//        if (file_exists($cache) and filemtime($cache) > strtotime('NOW - 1 WEEK')) {
-//            $ret =  json_decode(file_get_contents($cache), true);
-//            $ret['file'] = $cache . '.data';
-//            return $ret;
-//        }
+        if (file_exists($cache) and filemtime($cache) > strtotime('NOW - 1 WEEK')) {
+            $ret =  json_decode(file_get_contents($cache), true);
+            $ret['file'] = $cache . '.data';
+            return $ret;
+        }
         if (!file_exists(dirname($cache))) {
             mkdir(dirname($cache),0700, true);
         }
-        print_R($url);
-        echo "\n";
+        
         require_once 'HTTP/Request.php';
         $a = new HTTP_Request($url);
         $a->sendRequest();
-        print_R($a->getResponseBody());
-        echo "\n";
         file_put_contents($cache .'.data', $a->getResponseBody());
         
         $mt = $a->getResponseHeader('Content-Type');
